story_book_maker/pyproject.toml
[project]
name = "story-book-maker"
version = "0.1.0"
description = "AI 동화책 제작 에이전트"
requires-python = ">=3.10"
dependencies = [
    "google-adk>=1.27.0",
    "litellm>=1.75.5",
    "openai>=1.100.2",
    "pillow>=12.1.1",
]

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "story_book_maker"
module-root = ""
